Cadastro com previsão baseada na localização do usuário.

+++ app/src/main/java/com/jvmr/getclima/datasource/HGDataSource.java
@@ -21,7 +21,7 @@ public class HGDataSource {
         return instance;
     }
 
-    public CidadeModel buscarCidadePorGeoLoc(float latitude, float longitude) {
+    public CidadeModel buscarCidadePorGeoLoc(double latitude, double longitude) {
         CidadeModel cidadeModel = null;
         StringBuilder retorno;
